Fix crash on bad font settings
Some part of the font calculation code is throwing an exception that is
reaching the message loop and consequently terminating the VS process.
The most likely cause is the DTE.Properties object throwing an
exception.  This code handles the exception, logs and returns a default
value.

Going to install an easy to use system for users to get the trace
information from VsVim.  This is often very valuable and would allow us
to fully diagnose this problem from our side without forcing users to
attach a debugger

closes #1297