Fix mouse beep on selection
Selecting code with the mouse was causing a beep to occur.  The problem
was that mouse keys hadn't been suppressed in the CanProcess method of
several IMode implementations including Visual Mode.  This eventually
caused Process to be called and visual mode correctly beeped on the bad
key.  The fix is to return false in all modes that don't process mouse
key input.

Oddly enough really only normal mode processes mouse key input.  All
other modes either just react to selection events or see the commands
that mouse key input is mapped to.

closes #1317