Fixes exception in :wq
This stops the exception in the :wq command by correctly checking for
IsClosed in the OnEnter method of normal mode.  Long term I'm not sure
this is the best approach though.  Running any mode code after the
ITextView is closed just invites further errors like this one.  I'm
beginning to wonder if a better approach would be to have a ClosedMode
which we switch to whenever SwitchPreviousMode is called and the
ITextView is closed.  This is the simplest fix that can be done for the
point release though.

closes #1327