Ability to disable R# rename tracking with option
The options page now has the ability to disable R# + snippet rename
tracking.

In an ideal world this option would never be available.  However in the
past this code has been a perf issue and at the moment several users are
reporting performance problems.  Giving users the option to disable it
will allow for very fast diagnosis of whether this is the problem or not