Changed GetBackspacingPoint to provide more data
Previously the GetBackspacingPoint function would return simply the
point to which a backspace operation should delete.  In order to
implement 'softabstop' this function needs to provide more information.
A backspace with 'softtabstop' is very possibly a delete + an insert
operation.  Classic case is the following

1. Settings ts=8, sts=4, noet
2. Go to a blank line in insert mode and type <Tab><Tab>.  This will
result in a single \t character being inserted into the file
3. Type <BS> and the result will be 4 spaces on the line.

In that case 'sts' overrides 'noexpandtab' and splits the tab.  The only
logical way to represent this is as a replace operation.

This change simply updates the API to allow for the possibility of the
replace.  I wanted to separate out this change to make sure it didn't
interfere with the existing tests around 'backspace'