'sts' and backspace now working
The interaction between backspace and 'sts' is now working for the key
scenarios.

This change though did break the interaction with backspacing over the
start.  I need to rework that interaction interaction between InsertMode
and InsertUtil (in particular how it affects repeat and delete which
crosses start).  Temporarily disabled the tests.  Next check-in will be
re-enabling them and fixing the interaction