Fix the handling of 'backspace' and start option
This fixes the handling of the 'backspace' setting and the start option.

Unfortunately though this revealed a flaw in our edit repeat
architecture.  The edit repeat code depends on translating InsertCommand
operations into flat TextChange values.  In general this works fine but
not for backspace repeats.  The reason is that backspace actively
reexamines the 'backspace' setting every time it runs.  Hence the repeat
can't be flat text change, it's a semantic operation.  For now disabled
the involved tests, need to rethink this feature in light of the new
data.