Fixed KeyDown1 unit test failure
The KeyDown1 unit test was failing because we are unable to set the
keyboard layout to portuguese in AppVeyor.  This isn't actually
necessary to the test anymore though.  It was at one point in the past
when VsVim did interesting querying of keyboard layout.  It no longer
does that hence this feature can simply be removed