Fix issue with modal dialogs
The conditional breakpoint window now hosts an instance of the WPF
editor.  This was causing us to create a FallbackKeyProcessor for that
editor and it was turning key strokes like Ctrl+A into commands for the
active editor.  The fix here is rather straight forward, the fall back
key processor should not be operating in a modal dialog.

close #1380