Created separate extension
This change created a Dev14 only branch of the extension.  At the moment
there is a bug in the gallery which prevents an extension from
simultaneously having support for Visual Studio 2013 and Dev14.  As a
work around I'm temporarily supporting 2 extensions, one which is Dev14
only.

This branch represents that extension.  Once the gallery bug is fixed
this branch and that separate extension will be deleted