Changed the notification to loading of vimrc file
The impact of the 'backspace' setting is greatly diminished with the new
options approach.  A user must both have a vimrc file which changes the
setting, and switch the option to use it.  Hence having a notification
for it is over kill.

Instead this notification now warns about the automatic loading of the
vimrc file.  This is often quite unexpected by users.  Hopefully this
will remove the mystery a bit