Fixed majority of tests
Majority of tests working now.  The bulk of the problem was in the
RepeatLastCommand function.  It commonly opened nested undo and linked
undo transactions.

Still have to resolve one case.  What happens when you intentionally
create an empty linked transaction as is the case with a counted insert?
If the user types the sequence `3i<Esc>` then we have an empty linked
transaction.  Need to resolve this