Beginning of an IME fix
This is the very early stages of a fix for #1422.  It fixes the very
narrow scenario outlined in the bug but pretty much nothing else around
IME.  Going to need to put a bit of thought into this before it can be
called done.

A lot of work still needs to happen like figuring out how to deal with
buffered key inputs, insert and normal mode.