First step to :echo &expandtab (echo option)
Plumbed enough of the "echo option" stuff through to get a unit test
working without breaking any others. The type inference of the F#
compiler seems to be a bit of a doozy.

I'm not entirely comfortable yet with what I had to do here. It seems to
me that VariableValue has been used in a both of places in which
Expression should have been used. Still, I will definitely need
@JaredPar to review this.