Fix quoted string bug
The string content code wasn't properly considering full escape
sequences in string.  That requires left to right processing (or more
advanced right to left) but the code was just looking one character to
the left.  Changed the code to be more robust here with a full left to
right scan

closes #1454