Implemented :tabonly
This may not be the best way to have implemented it. I call
`RunVisualStudioCommand` in order to perform the work, and it feels a
little odd to add yet another layer of interpretation here. Thus, this
should probably be refactored. However, it does work.

Note that it doesn't work exactly the way it does in Vim due to the
impedance mismatch between Vim and Visual Studio's tab handling.
Specifically, because in Vim the splits exist within the tabs, whereas
in Visual Studio the tabs exist within the splits, in Vim if you have
multiple tabs and splits, `:tabonly` will leave you with the splits in
your current tab, but close all your other tabs, whereas with VsVim it
will close all your tabs in your current split, but leave you with the
tabs in your other splits.

Also, realized I had not followed the convention on unit tests names and
correct that wrong.