Added the explicit interface implementations
I missed this when I was reviewing the recent changes in this area.

I should probably elaborate on why I actually have this practice because
I'm sure it's not a popular one: in short I find that it is a code
maintenance win.

VsVim is a decently sized project at this time and it's easy for code to
get stale over time.  Before I used explicit interfaces I kept finding
dead code which used to implement interface methods (which had
themselves long ago been deleted).  I constantly found myself spending
considerable time trying to update algorithms and decipher old code only
to find that it was dead code all along.

Using explicit interfaces helps combat this.  If an interface changes
there is no way that it will silently impact an implementation now.  It
will force at least a casual glance at the implementation to see if it
is relevant or not.

Generally I don't care as much about this when an interface has 1 or 2
methods.  But past that I'm pretty rigid about explicit implementations.