Fixed up the unit tests
The only remaining failure was the memory leak tests.  It turns out the
test code was force creating CommandMargin which is an
IWpfTextViewMargin.  The WPF editor code path never asked for the
creation of the margin hence it never bothered to run Dispose and this
leads to a memory leak.