Fix double click selection issue
The double click selection behavior was incorrect in the latest build.
It was extending the selection by a single character.  This regression
was a result of the following commit

22fa32cf39c912053c9d725791b137122f974506

The intent of that commit was to have the selection extend to the caret
on drag operations (gVim) behavior.  It was triggered on double click
which caused that selection to also get extended by one character.

The fix is to make the drag selection logic more robust.  Actually check
for drag and not just left button click

This closes #1465