Keep VsVim out of Parallel Watch Window
The parallel watch window is labeling itself as a document and hence
VsVim is installing itself there.  This is a single line editor that
shouldn't be hosting a plugin like VsVim.

Long term I'm going to work with the parallel watch team to fix this
bug.  Short term add some logic to prevent VsVim from inserting itself
there.

closes #1490