Search forward start position fix
This fixes the position which a forward search occurs from.  Previously
the search would begin at the caret point + 1.  This allowed searches
like 'N' to function and I thought it was the correct behavior.

Experimentally though I found a different behavior.  If the text under
the caret matches the search then the search begins at the end of the
match.  The +1 behavior was partially correct and worked for a good deal
of patterns which is why this wasn't noticed until now

closes #1392