Add way to force remove TSQL bindings
There are a set of TSQL command entries which a) have no name and b) use
keystrokes
Vim users would like to use.  In general this isn't a problem becuase
VsVim is
capable of removing conflicting key bindings via the options page

The problem here is the commands are unnamed.  Any attempt to change a
key binding
on an unnamed command (via Command::put_Bindings) will fail.  There is
no good reason
for this that I was able to track down, it's just the behavior.

There is some hope this can be changed in 2015 or the next release but
in the
short term I need to use a different hammer for changing the key
bindings.  This method
method removes the binding by doing the following

1. Add a key binding to a VsVim command where the key binding has the
same scope as the
TSQL key bindings
2. Remove the key bindings on the VsVim command

As a consequence of adding the key binding in #1 Visual Studio will
delete the key
binding on the TSQL command.  This type of deletion doesn't check for a
name and hence
succeeds.  The subsequent removal of the key binding to the fake VsVim
command effectively
fully clears the key binding which is what we want