Implemented localtime() builtin VimL function
I chose localtime() not so much because it is particularly important,
but because I wanted to get a function with no arguments implemented to
enhance the parsing, and because its implementation was relatively
trivial, allowing me to focus on what I was really trying to do. At this
point, both localtime() and exists() are actually working if you run the
Vim App (I was not parsing the closing parenthesis before).