Implemented escape() builtin VimL function
This seemed liked a good multiple-argument case to start with. Have yet
to do an optional-argument case, but localtime(), exists(arg), and
escape(arg1 arg2) are all working. This is probably going to need some
refactoring, though: some things which are going on in the Interpreter
layer probably need to be pushed down into Vim.