Updated error messages per pull request discussion
Refactored some in the process. I realized that there were a number of
situations in which Vim will give "Using List as a string" or the like
as an error, including `:echo 'asdf' . []` and `:let @a=[]`, and so
decided to introduce a structure that supported that, because I was
noticing a pattern of having to convert numbers to strings for several
commands. This should also happen, strictly, in function calls, but
exists(1) (which returns 0) hardly seems valuable to me.