Implemented `:echo []` and `:echo {}`
I was trying to test the errors I put in in the last commit and couldn't
get them to come up, so I went ahead and improved the list and
dictionary support every so slightly so I could finish the testing.
Besides, it's probably something we would have wanted to do eventually
anyway.