Fix dd of collapsed region
Fix the 'dd' command when the caret is placed in the line break after a
collapsed region.  The code must remap the point back to the start of
the line before mapping up to the visual snapshot.  Otherwise mapping up
will fail because the start of the chosen line is collapsed and has no
mapping in the visual snapshot.

closes #1436