Errors from vimrc load recorded.
The logic now exists to capture errors during a vimrc load.
Unfortunately there is an underlying bug which is preventing anything
from actually being reported.

The IStatusUtil instance is only hooked up to the IVimBuffer fully once
the VimBufferCreated callback is issued.  That is not issued for the
vimrc buffer because it's specifically hidden from the view of the
extensibility mechanisms.  Even worse is the association is between
ITextBuffer and IStatusUtil when it should be for ITextView.

Need to clean up this entire area.