Deploy script now builds on all machines
The installed version of MSBuild sets a default value for
VisualStudioVersion.  This value is used to help determine the location
of the F# targets file.  This meant that VimCore could only find the
correct targets file if the version number chosen by MSBuild
corresponded with a version of Visual Studio installed on the machine.

The deploy script now just searches for Visual Studio versions on the
machine and overrides the default.  This means build will work just fine
so long as some supported version of Visual Studio is installed.