Post edit fix up invalidated snapshot
This is another case where a post edit fix up by a different component
invalidated an ITextSnapshot instance VsVim was using to position the
caret.

To counteract this I decided to make a systematic fix wit this change.
Added helper methods to ITextEdit and ITextBuffer for applying the
change and then getting the ITextSnapshot which returned after all edits
complete.  Searched through the code bases and changed all uses of this
API to the new helper methods.

closes #1554