Don't issue error during Roslyn rename
This adjusts the error detection logic in the UndoRedoOperations
implementation to not issue an error during a Roslyn Rename operation.
Roslyn manipulates the undo / redo stack directly during rename and by
default the UndoRedoOperations type sees this as an error (it wants to
control undo).  Implemented a host extension to give the host the
ability to state this was an intended operation.