Move ITextChangeTracker to Insert namespace
This component is only used for processing in Insert mode but was
available in the general Vim namespace.  On several occasions the
presence of this component in the Vim namespace has caused me to dig
into it to see if it could be removed or if there were communication
bugs between components that used it (it's a modal component).  Fixing
this by putting it in the Insert namespace where it belongs.