Corrected the broken tests
Unfortunately in this case the last remaining Skip test was checking
behavior that can't be implemented in VsVim.  The WPF editor can't
position the caret in the middle of a multi-space character like a tab.
It will instead always appear at the start.

Vim doesn't have this limitation when it is in a block selection.  The
caret is free to position on any space.

This deviation means we can't properly implement block replacements
where the caret needs to be positioned within the middle of a tab.  This
could be done if we wrote a separate caret implementation but for now
the need here doesn't rise to the complexity required to do that.