High contrast updates
This changes the command margin so that it will adjust to Visual Studio
being placed in high contrast mode.

This is not a strong area for me programming wise so it's hard to know
if this is a good solution, or a hack.  Testing it out though it appears
to work quite well.  So going with this for now and use it to get
feedback on how it is working.