BlockCaret after EOL - fix tests & AdjustCaretForVirtualEdit
CommonOperations.AdjustCaretForVirtualEdit - exclude any Visual kind of
modes. I looked at the behavior in vim, and it seems like with and
without ve=onemore setting. It appears to use the ve=onemore behavior.

Switch a few tests from [Fact] to [Theory] to test both ve="" and
ve="onemore" to use the public static property VirtualEditOptions on
VimTestBase.

InsertUtilTest - Set a few tests to Insert Mode before running the test
action. Most of the tests are OK just using the _insertUtilRaw. The
modified tests are relying on the additional behavior from
SelectionChangeTracker.

VisualModeIntegrationTests -
add SwitchEnterMode. This switches the mode, then sets the visual
selection (and does not run the check to see if _context.IsEmpty before
RunAll(). I don't think this is important for these tests.) This is only
used on tests that were previously failing when the change was made to the
SelectionChangeTracker
Modify a few tests to VisualLine mode before running the test action.