Fixed a Vs 2010 reference
The DLL MS.VS.Shell.Immutable.10.0 is used indirectly by the Platform APIs.  I made this an explicit reference.

Additionally the versions of the DLL shipped in newer Visual Studio instances has a difference TargetFramework value.  The VS 2010 version should always have a 4.0 TFM as that is the framework shipped in VS 2010.  But newer versions have 4.5 TFM which is incorrect.  Fixed this by adding the DLL to the reference set.