Don't track `virtualedit` on arbitrary mouse position changes
This essentially reverts a lot of the behavior in 4279e17676102eba70d182ab8bb2924b8534b58d.

When this change first came through I thought it was a good idea.  It seemed natural to properly enforce the `virtualedit` setting on any caret position change, not just when VsVim was processing key strokes.

Since this has shipped though customers have been sending a swarm of scenarios where this change broke behavior:

- #1789
- #1791
- #1792
- #1801

Fixing all these bugs is beginning to feel like a whack a mole problem.  There is simply no way to generalize the fix such that it excludes all of the places customers expect selection to hit EOL irrespective of the `virtualedit` setting.  As such going to revert back to the old behavior of not checking that on arbitrary mouse position changes.