Moved part of VsVim to 4.5 runtime
This is a necessary pre-req for adopting Application Insights.  It is supported on 4.0 but it brings along too many
dependencies with it for my comfort.  Once 4.5 is targeted it brings down a single DLL which is managable.

The only downside of moving to 4.0 is losing easy support for VS 2010.  That release only required 4.0 be installed.  But that is lost anyways by virtue of moving to VSIX V3.

This change also forced me to upgrade to a new versiont of Moq.  It itself not an issue but I did run into the following bug:

- https://github.com/moq/moq4/issues/265

Worked around it by moving Castle forward to a newer version.