Attempt work around for composition error
One customer is reporting a reliable composition error using latest
VsVim.  The root cause is a cycle issue with ITelemetry.  That is odd as
it doesn't show up for other users, but it is happening reliably.  Best
I can theorize is another extension on customer machine is causing a
slight change to ordering here that reveals a problem.

Going on this I'm going to try and work around this by breaking up the
dependencies a bit and turning it into a delayed create.

#1876