Proof of concept 'l' over surrogate pairs
There is an issue where if you try to navigate over a unicode surrogate
pair, you cannot move right over it. If you are before it, you will end up
in the character and get rounded back down to where you were. If you are
on the left of the character and try to move back to the right over the
character, you will end up in the character and get rounded back to the
intended position.

This fixes that in the case of insert mode, but the code is ugly and the
seam that I'm explointing isn't the most elegant. This serves as a good
base to start the fix for both normal and insert mode.