Fixed the deploy script
The deploy script now works with VS 2017.

For the time being the input was unfortunately changed to a VS
installation path vs. a version number.  This is due to the non-central
natirue of VS 2017.  There are APIs that can be used to find the
installs but didn't have the time to use them now.

Will come back to this later. There is nothing functionally wrong with
this approach, it just makes the script harder to invoke.