Fix crash on startup
When the editor raises the PerformLayout event it is not in a fully usable state.  Any operations within
that event which cause another layout to occur will result in an exception.  This breaks our delayed init
logic which is predicated on waiting till a time where the editor is fully usable.

Changed the delay init logic to do a simple Post inside the LayoutChanged handler.  Editor should be usable
then.

closes #1904