Fix recursive mode switch problem
The tests I added revealed a subtle bug in mode switches.  It was
possible to get into a case, at least in testing, where mode switches
layered on top of mode switches.  This should not happen as it's not
really possible to program against.

Changed the code to throw when this happens and fixed the underlying
bug.