Move to a hard reference model
The test apps and dlls now reference the actual binaries that they
run against. This is necessary given that the 2017 binaries are no
longer in the GAC. The only way to run these projects is to have
all of the right binaries in the output directory.

Bonus is that it works fine with old VS versions where everything
is in the GAC.

This will make it harder to test older versions (prior to 2015)
because I dont have those ref binaries checked in. But if any
issues come up I can just track them down.