Fixed how background tasks are scheduled
The default scheduler for Task.Start is the current TaskScheduler
instance. That is typically the UI scheduler. Hence all of the work
which was designed to be running on the background thread was actually
running on the UI thread instead. Oops ....

Use TaskScheduler.Default explicitly to ensure that the work is
scheduled on the thread pool.