Moved to a single build step
Previously was building a lot of projects individually. Can't really say
for certain what I was thinking here in the past. Most likely it had to
due with the repo being structured around a lot of small solution files.

Now though there is an all encompassing solution. Using this one allows
a single build to happen.