Moved Deploy under binaries
A core principal of builds is they should never mutate the source
directory. For most projects there is an exception granted for a root
directory that holds all build output. For exampel Binaries

VsVim was violating this by having the deploy script create and populate
a seperate root directory called Deploy. Changed that to be a
sub-directory of the Binaries directory.