Clear status bar when we set the text
The status bar should only be cleared when it contains the text that
VsVim had previously set. This is a shared resource and it's possible
other components have added text here. In the case we are displaying
nothing we should only clear it if the current text is what we
previously set.

closes #2028