Got a simple menu displaying
The problem with the existing menu is all of the commands inside
VsVimMenuGroup were hidden by virtue of being inside the command well. A
menu won't display unless there is one visible item in the group.

The new menu is working because there are two visible commands inside of
it. This commands display as a part of the same "group". That is they
are grouped together between little display bars under the "Tools" menu.