Fix race condition on startup
The behavior here is that we get an `ITextView.LayoutChanged` event but
by the time we can react the `ITextView` is already closed. The fix is to
move our contract check that the `ITextView` is in a proper state to
after we've checked to see if it has been closed.

closes #2250