Removed List allocations in KeyInputSet
The design of `KeyInputSet` was meant to avoid allocating
`FSharpList<KeyInput>` instances. Instead choosing a discriminated union
to have efficient allocation for the common case of one or two elements.

This worked for the initial `KeyInputSet` allocation but the
implementation of several functions, most notably `CompareTo` ended up
allocating `FSharpList<KeyInput>`. These were not stored and hence
allocated repeatedly on every call instead of once up front.

I made a lot of attempts at refactoring these methods to not have the
allocations but the code was awkward and I ended up chasing down other
allocations that got added. Eventually I decided to abandon this
approach.

Profiling showed that this allocation dominates large sections of our
unit test runs. It represented ~15% of all allocations and was a
significant outlier.

In looking at the use cases of `KeyInputSet` I also determined that the
idea that a `OneKeyInput` could be equal to a `ManyKeyInputs` wasn't
always respected. This further made me question the design of the API.

After consideration decided to allocate a list up front and use that.
This overall cut down on the allocations considerably.