Terminology update
This changes the terminology of CharacterSpan and a few other types to
be more reflective of their actual behavior. One item in particular is
understanding that in CharacterSpan we are typically tracking the
maximum, or rather ideal, length of the last line. The actual Span
occupied by the CharacterSpan instance may not include that position.
Yet it's important to preserve as the CharacterSpan moves around within
the buffer in case it can expand in a new location.