API cleanup: use GetFor instead of CreateFor
The `Get` pattern seemed to be the dominant choice in the new APIs I wrote. Decided on that over `Create` on that basis (and no concrete reasons). Overall though I'm unhappy with the `Create` based methods in the code base. I should have embraced constructors and optional parameters a bit better.