Fix Roslyn refactor
The Roslyn refactoring types moved between assemblies in recent VS
releases. Need to account for that when searching for the types that are
used.

Note there will still be one "Undo / Redo Unexpected" warning when
ininiating a refactor rename operation. This is really only visible when
run under the debugger though and it's due to an operation very early in
the rename startup sequence: before it's possible to detect a rename.

Looked at this for a while to see if there was a way to detect this
properly but could not find a case. Given it's not really observable and
refactor is working again we're okay for now. Will follow up with Roslyn
team to get a better way to detect this.

close #2088