Fix surrogate pair block insert
This addresses block insert around surrogate pairs. Particularly when
the block insert column is in the middle of a wide surrogate pair. This
should be done by pretending the column was the start of the surrogate
pair. This is different than wide characters which are not surrogate
pairs; those are converted to spaces first and then block insert occurs
in those spaces.

closes #2344