Fix MEF composition error
This fixes a MEF composition error that occurs more frequently on older
versions of Visual Studio. There appears to be a timing issue with
accessing the output window on the UI thread during composition.

Unfortunately I'm not able to see what is going wrong as it's happening
within the DTE layer. The fix though is pretty straight forward:

1. Delay the initialization of the Output Window to idle
1. Catch exceptions coming from that code and raise errors vs. letting
them be unhandled

closes #2249