Remove ISmartTagBroker
This type is deleted in Dev16 which means VsVim can't load as long as it
continues to use the type in the core DLLs. I considered moving the
usage to the VS specific DLLs. However upon further inspection I don't
think the usage is needed anymore. Most of the logic is ensuring we
allow arrow keys to pass through to the menu for smart tags / ligt
bulbs. That happens even with all of the smart tag logic disabled.

Given that just deleting the implementation for now. Can revisit with a
VS specific approach if we end up finding the need for it.