Change SynchronizationContext in testing
Need to start off with a bit of a primer for how
`SynchronizationContext` work in an xunit environment. The xunit
framework monitors all calls to the active `SynchronizationContext`
during the execution of a test. It will insert code that essentially
waits on these calls, `Post` and `Send`, to complete before the test
which created them can complete.

Methods which are not `async` are suspectible to this happening as they
have no `Dispatcher` running that can clear out the posted messages.
Unless they take explicit action to clear out the
`SynchronizationContext` they will deadlock / fail whenever a message
gets posted. This means only synchronous methods which expect messages
to be posted + properly pump them will execute correctly. Changes which
unexpectedly introduce posted messages will end up chasing down a number
of hung tests.

Initially I added `TestableSynchronizationContext` as a solution to this
problem since virtually 100% of the test methods are synchronous. It had
some detection for the bad case here: This solution has a couple of
failings:

1. It can't detect the unexpected posting as they happen. Instead it
attempts to detect them at the end of test execution.
1. It's not compatible with `async` test methods and given it's
installed by default it effectively means you can't use `async` test
methods in the code base.

The new solution attempts to address these problems. The
`VimSynchronizationContext` just intercepts all `Send` and `Post` calls
and throws immediately if they're unexpected.

The need for `async` test methods became a bit of a necessity
while working on #2465