Fix load error in C# script executor
There are work loads of Visual Studio that don't have the C# scripting
assemblies installed. For example C++ and Python can do this.

For such work loads the C# script execution functionality won't load
correctly. Need to guard against this case and fall back to not
supported behavior when it happens.

closes #2500