Don't fade completion
The default editor behavior is to fade the completion window when the
control key is pressed. That is unfortunate because that's the default
key binding for vim based completion behavior.

This change implements a work around by suppressing this behavior. The
reflection here is a necessary evil as there is no public way to prevent
this from happening.