Shutdown Dispatcher properly in tests
This change uses the `ITestFramework` hook in xUnit to allow for
graceful shutdown of the `Dispatcher` instance used to run our tests.
The `ITestFramework.Dispose` method is both

1. Called before `AppDomain.Unload`
1. Called after all tests have completed executing

This means it's the ideal place to tear down the `Dispatcher` and STA
thread created to run the tests.

Other places considered for tear down:

1. `AppDomain.CurrentDomain.DomainUnload`: this event fires as the
`AppDomain` is being torn down meaning the `Dispatcher` or its child
objects could have already been finalized. Indeed this is the case on
most of the machines tested. This meant calling
`Dispatcher.BeginInvokeShutdown` was unreliable as the underlying HWND
handles may have been finalized.
1. `StaContext.Dispose`: this is not a part of any hook in xUnit hence
is never actually invoked.

Note: this may end up breaking source based discovery for tests in Test
Explorer but normal build based discovery will continue working.

closes #2563