Clean up build files
This cleans up the scripts used to build, test and create releases for
VsVim. This is mostly just a rename of Delpoy.ps1 + clean up to support
testing on VS2017 and VS2019.

This is generally how I create VsVim releases. Deploy.ps1 is presently
broken (my fault) and it needs to be fixed up before releasing 2.7.0.

This change is factored out of PR #2575 and you can see the validation
of it in that PR.