Fix the window display
Changed the VS Integration test to display the editor during some unit
tests. In order to properly test the scenario the window must be
`IsVisible == true`. Lacking that items like the intellisense builders
will dismiss the relevant sessions thinking the control has been hidden.

Truly wanted to avoid displaying UI during these tests but don't see a
way out of it. The intellisense primitives are hard coded to dismiss
items when they're not visible. The editor team takes this approach for
the same reasons.