Publish should only occur on master
VSIX gallery publishing should only occur on builds in the master
branch.

Before this change publish was occuring whenever the build was
IndividualCI or BatchedCI. The CI term in Azure DevOps means a build
that occurs on merged code (not in PR). This meant every time DevOps was
triggered on merged code the result was published.

The initial intent was that only 'master' was setup to build on merge
but some developmental code got left in and it ended up triggering on
dev branches as well as master. This fixes both problems.