Move to xunit 2.4.1
Profiling the unit tests today I noticed we were still on xunit 2.3.1
which is nearly two years old. Decided to move us to a more recent
version.

This change also removes all of our checked in xunit binaries. The code
now relies completely on the binaries from the NuPkg files. This will
make upgrading easier in the future.