Fix UI delay in navigate to
The `UnwantedSelectionHandler` is properly using the
`_VSRDTFLAGS4.RDT_PendingInitialization` when iterating the RDT for open
documents. It seems though that this flag is not respected for project
files and forces them to be realized by Visual Studio. Loading a project
file can be fairly expensive and this gets done synchronously when
begining a navigate to operation.

Long term this does seem like a bug that Visual Studio should look into.
Short term though the `UnwantedSelectionHandler` can avoid the RDT
entirely by just walking through `IVim.VimBuffers` instead. This
represents the active set of `IVimBuffer` instances and that is
where we want to control the selection before / after a
navigate to operation.

closes 2715