Remove VimGrep functionality
Several classes in VSMac became internal at some point. These were
related to "Find in Files" functionality.

This unfortunately means that we need to remove VimGrep functionality for now.
We can probably reinstate it by providing the UI for VimGrep ourselves. It's
not a good time to do this right now though as VSMac is transitioning from GTK
to use Cocoa for UI.

Update VSMac and Mono to latest stable versions.