Moved towards PackageReference
Got the VS2019 build to use `PackageReference` but had a lot of problems
with the VS2017 builds. The mix of `net45` and `net472` target
frameworks were causing a lot of issues with how
`Microsoft.VisualStudio.Threading` was getting resolved. Couldn't really
understand how it was happening.

Decided to effectively downgrade the VS2017 to 2015 references. Get that
working and get to a known good, commit and then try a completely
separate commit to get it working on 2017 again