Made progress on MS.VS.Threading issue
The problem turned out to be logic in the app.config files of the VSIX
projects. These had binding redirect logic in it for several DLLs and
that, apparently, was contributing to how `ResolveAssemblyReference`
functions. I'm unsure how but once I removed the app.config files (which
weren't needed) the issues went away.

Something to dig into later to see what actually was happening here.
Good news though is I'm back on track. Need to undo the part where I
moved the VS2017 project no 2015 but that (hopefully) can make progress
now.